requirements.txt
fastapi
uvicorn
tortoise-orm
python-multipart
python-jose[cryptography]
passlib[bcrypt] 